Initial work migrating demo app to real components
