+++ lib/field-array-item/index.ts
@@ -1,0 +1,2 @@
+export * from "./field-array-item";
+export * from "./types";
