+++ lib/field-array-item/types.ts
@@ -1,0 +1,3 @@
+import { FieldInstance } from "../field/types";
+
+interface FieldArrayItemInstance<T = any> extends FieldInstance<T> {}
