+++ lib/field-array/field-array.spec.tsx
@@ -1,0 +1,7 @@
+import { test } from "vitest";
+
+test.todo("should register a field array with the form");
+test.todo(
+  "field array should allow you to set an initial value for the array of values"
+);
+test.todo("field array should track `isDirty` for the array of values");
