+++ lib/index.ts
@@ -1,3 +1,5 @@
 export * from "./field";
 export * from "./form";
 export * from "./utils";
+export * from "./field-array";
+export * from "./field-array-item";
